+++ cypress/locators/action-popover/index.js
@@ -2,6 +2,7 @@ import {
   ACTION_POPOVER_BUTTON,
   ACTION_POPOVER_DATA_COMPONENT,
   ACTION_POPOVER_DIVIDER,
+  ACTION_POPOVER_SUBMENU,
 } from './locators';
 
 // component preview locators
@@ -9,5 +10,11 @@ export const actionPopoverButton = () => cy.get(ACTION_POPOVER_BUTTON);
 export const actionPopoverButtonNoIframe = () => cy.iFrame(ACTION_POPOVER_BUTTON);
 export const actionPopover = () => cy.get(ACTION_POPOVER_DATA_COMPONENT);
 export const actionPopoverDivider = () => cy.get(ACTION_POPOVER_DIVIDER);
-export const actionPopoverInnerItem = index => cy.iFrame(ACTION_POPOVER_DATA_COMPONENT).eq(0)
-  .find(`div:nth-child(${index})`);
+export const actionPopoverInnerItem = index => cy.iFrame(ACTION_POPOVER_DATA_COMPONENT).first()
+  .children().eq(index);
+export const actionPopoverSubmenu = index => cy.iFrame(ACTION_POPOVER_SUBMENU)
+  .eq(1).children().eq(index);
+export const actionPopoverSubmenuInnerElementNoIFrame = index => cy.get(ACTION_POPOVER_DATA_COMPONENT)
+  .find('nth:child(3)')
+  .children().eq(index);
+export const actionPopoverSubmenuNoIFrame = () => cy.get(ACTION_POPOVER_SUBMENU).eq(1);