+++ cypress/support/step-definitions/profile-steps.js
@@ -14,6 +14,18 @@ Then('avatar is taken from {string}', (avatarUrl) => {
   avatarPreview().should('have.attr', 'src', avatarUrl);
 });
 
+Then('Profile is set to large', () => {
+  profile().should('have.class', PROFILE_LARGE);
+});
+
+Then('name is set to {string}', (name) => {
+  namePreview().should('have.text', name);
+});
+
+Then('Profile size has {string}', (property) => {
+  initialsPreview().should('have.css', 'height', `${property}px`);
+});
+
 Then('initials is set to {string}', (initials) => {
   // eslint-disable-next-line no-param-reassign
   initials = initials.substring(0, 3);
@@ -21,12 +33,4 @@ Then('initials is set to {string}', (initials) => {
     initialsPreview().children()
       .should('have.attr', 'src', `${DATA_IMAGE_PREFIX}${$initials}`);
   });
-});
-
-Then('Profile is set to large', () => {
-  profile().should('have.class', PROFILE_LARGE);
-});
-
-Then('name is set to {string}', (name) => {
-  namePreview().should('have.text', name);
 });