+++ cypress/support/step-definitions/textbox-steps.js
@@ -3,7 +3,7 @@ import {
 } from '../../locators/textbox';
 import {
   label, labelByPosition, commonDataElementInputPreview, commonInputPreview,
-  fieldHelpPreviewByPosition, tooltipPreviewByPosition,
+  fieldHelpPreviewByPosition, tooltipPreviewByPosition, labelNoIFrame, commonDataElementInputPreviewNoIframe,
 } from '../../locators';
 import { DEBUG_FLAG } from '..';
 
@@ -243,4 +243,22 @@ When('I click on Textbox', () => {
 
 Then('Textbox input has golden border on focus', () => {
   textbox().should('have.css', 'outline', 'rgb(255, 181, 0) solid 3px');
+});
+
+Then('Textbox overriden styles rendered properly', () => {
+  labelNoIFrame().parent().parent().should('have.css', 'background', 'rgb(240, 240, 240) none repeat scroll 0% 0% / auto padding-box border-box');
+  labelNoIFrame().parent().should('have.css', 'display', 'flex');
+  labelNoIFrame().should('have.css', 'display', 'block')
+    .and('have.css', 'font-weight', '600')
+    .and('have.css', 'box-sizing', 'border-box')
+    .and('have.css', 'padding-bottom', '0px')
+    .and('have.css', 'padding-top', '12px')
+    .and('have.css', 'padding-right', '11px')
+    .and('have.css', 'color', 'rgb(180, 212, 85)');
+  commonDataElementInputPreviewNoIframe().parent().should('have.css', 'background', 'rgb(255, 255, 255) none repeat scroll 0% 0% / auto padding-box border-box')
+    .and('have.css', 'border', '1px solid rgb(102, 132, 145)')
+    .and('have.css', 'flex', '0 0 auto');
+  commonDataElementInputPreviewNoIframe().should('have.css', 'background', 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box')
+    .and('have.css', 'border', '0px none rgba(0, 0, 0, 0.9)')
+    .and('have.css', 'color', 'rgba(0, 0, 0, 0.9)');
 });