+++ src/__experimental__/components/simple-color-picker/color-sample-box/color-sample-box.style.js
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 import PropTypes from 'prop-types';
 import colorSampleBoxClassicStyle from './color-sample-box-classic.style';
 
@@ -11,6 +11,14 @@ const StyledColorSampleBox = styled.div`
   justify-content: center;
   border: 2px solid transparent;
   background-color: ${({ color }) => color};
+
+  ${({ color }) => color === 'transparent' && css`
+    background: conic-gradient(#fff 0deg ,#fff 90deg,
+                grey 90deg,grey 180deg,
+                #fff 180deg,#fff 270deg,
+                grey 270deg,grey 360deg)
+                0 0/25% 25%;
+  `}
 
   ${colorSampleBoxClassicStyle}
 `;