+++ src/components/duelling-picklist/picklist-divider.component.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { StyledPicklistDivider } from './duelling-picklist.style';
+
+const PicklistDivider = props => <StyledPicklistDivider { ...props } data-element='picklist-divider' />;
+
+export default PicklistDivider;